requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile
#
annotated-types==0.7.0
    # via pydantic
cffi==1.17.1
    # via cryptography
charset-normalizer==3.4.2
    # via pdfminer-six
click==8.2.1
    # via -r requirements.in
colorama==0.4.6
    # via click
cryptography==45.0.4
    # via pdfminer-six
jinja2==3.1.6
    # via -r requirements.in
markupsafe==3.0.2
    # via jinja2
numpy==2.3.1
    # via -r requirements.in
pdfminer-six==20250506
    # via pdfplumber
pdfplumber==0.11.7
    # via -r requirements.in
pillow==11.2.1
    # via
    #   -r requirements.in
    #   pdfplumber
pycparser==2.22
    # via cffi
pydantic==2.11.7
    # via
    #   -r requirements.in
    #   pydantic-yaml
pydantic-core==2.33.2
    # via pydantic
pydantic-yaml==1.5.1
    # via -r requirements.in
pypdfium2==4.30.1
    # via pdfplumber
ruamel-yaml==0.18.14
    # via pydantic-yaml
ruamel-yaml-clib==0.2.12
    # via ruamel-yaml
typing-extensions==4.14.0
    # via
    #   pydantic
    #   pydantic-core
    #   pydantic-yaml
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
unidecode==1.4.0
    # via -r requirements.in
watchdog==6.0.0
    # via -r requirements.in
